removed portfolioSlice circ deps
