+++ src/hooks/useBalanceChartData/useBalanceChartData.ts
@@ -19,7 +19,10 @@ import { useFetchPriceHistories } from 'hooks/useFetchPriceHistories/useFetchPri
 import { bn, bnOrZero } from 'lib/bignumber/bignumber'
 import { AccountSpecifier } from 'state/slices/accountSpecifiersSlice/accountSpecifiersSlice'
 import { PriceHistoryData } from 'state/slices/marketDataSlice/marketDataSlice'
-import { PortfolioAssets, PortfolioBalancesById } from 'state/slices/portfolioSlice/portfolioSlice'
+import {
+  PortfolioAssets,
+  PortfolioBalancesById
+} from 'state/slices/portfolioSlice/portfolioSliceTypes'
 import {
   selectPortfolioAssets,
   selectPortfolioCryptoBalancesByAccountId,
