+++ src/state/slices/portfolioSlice/selectors.ts
@@ -17,7 +17,7 @@ import {
   PortfolioAssetBalances,
   PortfolioAssets,
   PortfolioBalancesById
-} from './portfolioSlice'
+} from './portfolioSliceTypes'
 import { accountIdToFeeAssetId, findAccountsByAssetId } from './utils'
 
 // We should prob change this once we add more chains
